+++ distage/distage-model/src/main/scala/com/github/pshirshov/izumi/distage/model/reflection/universe/DIUniverse.scala
@@ -3,10 +3,11 @@ package com.github.pshirshov.izumi.distage.model.reflection.universe
 import com.github.pshirshov.izumi.distage.model.definition.WithDIAnnotation
 import com.github.pshirshov.izumi.distage.model.plan.{WithDIAssociation, WithDIDependencyContext, WithDIWiring}
 import com.github.pshirshov.izumi.distage.model.references._
+import com.github.pshirshov.izumi.fundamentals.reflection.WithTags
 
 trait DIUniverse
   extends DIUniverseBase
-    with WithDITypeTags
+    with WithTags
     with WithDISafeType
     with WithDISymbolInfo
     with WithDITypedRef
