+++ fundamentals/fundamentals-reflection/src/main/scala/com/github/pshirshov/izumi/fundamentals/reflection/MacroTrivialSink.scala
@@ -4,7 +4,6 @@ import com.github.pshirshov.izumi.fundamentals.platform.console.AbstractStringTr
 
 import scala.reflect.macros.blackbox
 
-
 class MacroTrivialSink(c: blackbox.Context) extends AbstractStringTrivialSink {
   override def flush(value: => String): Unit =
     c.info(c.enclosingPosition, value, force = true)
